Implement most of the networking stuff
Move the dependency checking to immediately after execution
Make Loggable simpler by moving the caller formatting to Formatter

It should be noted that for now, cool.io doesn't support local IP
binding. I hacked it up to do so, but we'll see what happens between
now and when it comes time to get serious.

The main loop implementation is proof-of-concept at the moment. Things
have become sufficiently complex that I need to rework the way the
Kythera module works on a basic scale. It's probably going to have to
be a class. I'm not sure yet, but this is the next task.