Get rid of Message and just go back to origin, parv
There was just no benefit in using a class for this. Nothing was using
the extra data provided by the class, and that functionality ends up
being provided by User or Server. Maybe I'll replace origin with either
a User or Server depending on the context. I can do that in TS6 because
SIDs start with a number, but in other protocols it's not so easy.